+++ test/unit/ChildrenAwareNodeInterfaceTest.php
@@ -1,34 +1,36 @@
 <?php
 
-namespace Dhii\Data\Tree\FuncTest;
+namespace Dhii\Data\Tree\UnitTest;
 
 use Xpmock\TestCase;
 
 /**
- * Tests {@see \Dhii\Data\Tree\NodeInterface}.
+ * Tests {@see \Dhii\Data\Tree\ChildrenAwareNodeInterface}.
  *
  * @since [*next-version*]
  */
-class NodeInterfaceTest extends TestCase
+class ChildrenAwareNodeInterfaceTest extends TestCase
 {
     /**
      * The name of the test subject.
      *
      * @since [*next-version*]
      */
-    const TEST_SUBJECT_CLASSNAME = 'Dhii\\Data\\Tree\\NodeInterface';
+    const TEST_SUBJECT_CLASSNAME = 'Dhii\Data\Tree\ChildrenAwareNodeInterface';
 
     /**
      * Creates a new instance of the test subject.
      *
      * @since [*next-version*]
      *
-     * @return \Dhii\Data\Tree\NodeInterface
+     * @return \Dhii\Data\Tree\ChildrenAwareNodeInterface
      */
     public function createInstance()
     {
         $mock = $this->mock(static::TEST_SUBJECT_CLASSNAME)
             ->getValue()
+            ->getChildren()
+            ->hasChildren()
             ->new();
 
         return $mock;
@@ -43,11 +45,7 @@ class NodeInterfaceTest extends TestCase
     {
         $subject = $this->createInstance();
 
-        $this->assertInstanceOf(
-            static::TEST_SUBJECT_CLASSNAME, $subject, 'A valid instance of the subject could not be created'
-        );
-        $this->assertInstanceOf(
-            'Dhii\\Data\\ValueAwareInterface', $subject, 'Subject does not extend required interface'
-        );
+        $this->assertInstanceOf(static::TEST_SUBJECT_CLASSNAME, $subject, 'A valid instance of the subject could not be created');
+        $this->assertInstanceOf('Dhii\\Data\\Tree\\NodeInterface', $subject, 'Subject does not extend required interface');
     }
 }
