+++ test/unit/KeyAwareNodeInterfaceTest.php
@@ -1,36 +1,35 @@
 <?php
 
-namespace Dhii\Data\Tree\FuncTest;
+namespace Dhii\Data\Tree\UnitTest;
 
 use Xpmock\TestCase;
 
 /**
- * Tests {@see \Dhii\Data\Tree\ChildrenAwareNodeInterface}.
+ * Tests {@see \Dhii\Data\Tree\KeyAwareNodeInterface}.
  *
  * @since [*next-version*]
  */
-class ChildrenAwareNodeInterfaceTest extends TestCase
+class KeyAwareNodeInterfaceTest extends TestCase
 {
     /**
      * The name of the test subject.
      *
      * @since [*next-version*]
      */
-    const TEST_SUBJECT_CLASSNAME = 'Dhii\\Data\\Tree\\ChildrenAwareNodeInterface';
+    const TEST_SUBJECT_CLASSNAME = 'Dhii\Data\Tree\KeyAwareNodeInterface';
 
     /**
      * Creates a new instance of the test subject.
      *
      * @since [*next-version*]
      *
-     * @return \Dhii\Data\Tree\ChildrenAwareNodeInterface
+     * @return \Dhii\Data\Tree\KeyAwareNodeInterface
      */
     public function createInstance()
     {
         $mock = $this->mock(static::TEST_SUBJECT_CLASSNAME)
             ->getValue()
-            ->getChildren()
-            ->hasChildren()
+            ->getKey()
             ->new();
 
         return $mock;
@@ -45,11 +44,7 @@ class ChildrenAwareNodeInterfaceTest extends TestCase
     {
         $subject = $this->createInstance();
 
-        $this->assertInstanceOf(
-            static::TEST_SUBJECT_CLASSNAME, $subject, 'A valid instance of the subject could not be created'
-        );
-        $this->assertInstanceOf(
-            'Dhii\\Data\\Tree\\NodeInterface', $subject, 'Subject does not extend required interface'
-        );
+        $this->assertInstanceOf(static::TEST_SUBJECT_CLASSNAME, $subject, 'A valid instance of the subject could not be created');
+        $this->assertInstanceOf('Dhii\\Data\\Tree\\NodeInterface', $subject, 'Subject does not extend required interface');
     }
 }
