+++ test/unit/NodeInterfaceTest.php
@@ -1,35 +1,34 @@
 <?php
 
-namespace Dhii\Data\Tree\FuncTest;
+namespace Dhii\Data\Tree\UnitTest;
 
 use Xpmock\TestCase;
 
 /**
- * Tests {@see \Dhii\Data\Tree\KeyAwareNodeInterface}.
+ * Tests {@see \Dhii\Data\Tree\NodeInterface}.
  *
  * @since [*next-version*]
  */
-class KeyAwareNodeInterfaceTest extends TestCase
+class NodeInterfaceTest extends TestCase
 {
     /**
      * The name of the test subject.
      *
      * @since [*next-version*]
      */
-    const TEST_SUBJECT_CLASSNAME = 'Dhii\\Data\\Tree\\KeyAwareNodeInterface';
+    const TEST_SUBJECT_CLASSNAME = 'Dhii\Data\Tree\NodeInterface';
 
     /**
      * Creates a new instance of the test subject.
      *
      * @since [*next-version*]
      *
-     * @return \Dhii\Data\Tree\KeyAwareNodeInterface
+     * @return \Dhii\Data\Tree\NodeInterface
      */
     public function createInstance()
     {
         $mock = $this->mock(static::TEST_SUBJECT_CLASSNAME)
             ->getValue()
-            ->getKey()
             ->new();
 
         return $mock;
@@ -44,11 +43,7 @@ class KeyAwareNodeInterfaceTest extends TestCase
     {
         $subject = $this->createInstance();
 
-        $this->assertInstanceOf(
-            static::TEST_SUBJECT_CLASSNAME, $subject, 'A valid instance of the subject could not be created'
-        );
-        $this->assertInstanceOf(
-            'Dhii\\Data\\Tree\\NodeInterface', $subject, 'Subject does not extend required interface'
-        );
+        $this->assertInstanceOf(static::TEST_SUBJECT_CLASSNAME, $subject, 'A valid instance of the subject could not be created');
+        $this->assertInstanceOf('Dhii\\Data\\ValueAwareInterface', $subject, 'Subject does not extend required interface');
     }
 }
